NewsTeasersTrait [skip ci]
