+++ web/modules/custom/server_general/src/InnerElementTrait.php
@@ -33,89 +33,6 @@ trait InnerElementTrait {
   protected $renderer;
 
   /**
-   * Build "Card with image" for News content type.
-   *
-   * @param array $image
-   *   The image render array.
-   * @param string $title
-   *   The title.
-   * @param \Drupal\Core\Url $url
-   *   The URL to link to.
-   * @param array $summary
-   *   Summary of the search result.
-   * @param int $timestamp
-   *   The timestamp.
-   *
-   * @return array
-   *   Render array.
-   */
-  protected function buildInnerElementWithImageForNews(array $image, string $title, Url $url, array $summary, int $timestamp): array {
-    $elements = [];
-
-    // Labels.
-    $element = $this->buildLabelsFromText([$this->t('News')]);
-    $elements[] = $this->wrapTextResponsiveFontSize($element, 'sm');
-
-    // Date.
-    $element = IntlDate::formatPattern($timestamp, 'short');
-    $element = $this->wrapTextColor($element, 'gray');
-    $elements[] = $this->wrapTextResponsiveFontSize($element, 'sm');
-
-    // Title as link.
-    $element = $this->buildLink($title, $url, 'dark-gray');
-    $element = $this->wrapTextResponsiveFontSize($element, 'lg');
-    $elements[] = $this->wrapTextFontWeight($element, 'bold');
-
-    // Body teaser.
-    $elements[] = $this->wrapTextLineClamp($summary, 4);
-
-    return $this->buildInnerElementLayoutWithImage($url, $image, $elements);
-  }
-
-  /**
-   * Build "Card with image horizontal" for News content type.
-   *
-   * @param array $image
-   *   The image render array.
-   * @param string $title
-   *   The title.
-   * @param \Drupal\Core\Url $url
-   *   The URL to link to.
-   * @param array $summary
-   *   Summary of the search result.
-   * @param int $timestamp
-   *   The timestamp.
-   *
-   * @return array
-   *   Render array.
-   */
-  protected function buildInnerElementWithImageHorizontalForNews(array $image, string $title, Url $url, array $summary, int $timestamp): array {
-    $elements = [];
-
-    // Labels.
-    $element = $this->buildLabelsFromText([$this->t('News')]);
-    $elements[] = $this->wrapTextResponsiveFontSize($element, 'sm');
-
-    // Date.
-    $element = ['#markup' => IntlDate::formatPattern($timestamp, 'short')];
-    $element = $this->wrapTextColor($element, 'gray');
-    $elements[] = $this->wrapTextResponsiveFontSize($element, 'sm');
-
-    // Title as link.
-    $element = $this->buildLink($title, $url, 'dark-gray');
-    $element = $this->wrapTextResponsiveFontSize($element, 'lg');
-    $elements[] = $this->wrapTextFontWeight($element, 'bold');
-
-    // Body teaser.
-    $elements[] = $this->wrapTextLineClamp($summary, 4);
-
-    // Read more button.
-    $elements[] = $this->buildButton($this->t('Read more'), $url);
-
-    return $this->buildInnerElementLayoutWithImageHorizontal($url, $image, $elements);
-  }
-
-  /**
    * Build "Search result".
    *
    * @param string $label
